refactor: reorganize queries directory with domain-based structure
Reorganize src/queries from flat structure to domain-based organization
for improved maintainability and scalability.

Changes:
- Group operations by domain (tasks/, streams/, user/)
- Consolidate 9 task mutations into single mutations.ts file
- Rename fragments/mutations.ts to mutation-responses.ts for clarity
- Maintain backward compatibility through barrel exports

Benefits:
- Reduces code duplication (~1,600 lines removed)
- Improves discoverability (all task ops in tasks/)
- Scales better as new resources are added
- No breaking changes to public API

All tests passing (154 tests), TypeScript compilation successful.
