+++ src/queries/fragments/index.ts
@@ -4,3 +4,4 @@
 
 export * from './task.js';
 export * from './stream.js';
+export * from './mutation-responses.js';
