+++ src/queries/index.ts
@@ -2,8 +2,7 @@
  * GraphQL queries and fragments exports
  */
 
-export * from './user.js';
-export * from './tasks.js';
-export * from './streams.js';
+export * from './tasks/index.js';
+export * from './streams/index.js';
+export * from './user/index.js';
 export * from './fragments/index.js';
-export * from './mutations/index.js';
