--- src/queries/mutations/index.ts
@@ -1,13 +1,0 @@
-/**
- * GraphQL mutations exports
- */
-
-export * from './updateTaskComplete.js';
-export * from './updateTaskDelete.js';
-export * from './createTask.js';
-export * from './updateTaskSnoozeDate.js';
-export * from './updateTaskNotes.js';
-export * from './updateTaskPlannedTime.js';
-export * from './updateTaskDueDate.js';
-export * from './updateTaskText.js';
-export * from './updateTaskStream.js';
