+++ src/queries/streams/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Stream-related GraphQL operations exports
+ */
+
+export * from './queries.js';
