+++ src/queries/streams/queries.ts
@@ -3,7 +3,7 @@
  */
 
 import { gql } from 'graphql-tag';
-import { STREAM_FRAGMENT } from './fragments/index.js';
+import { STREAM_FRAGMENT } from '../fragments/index.js';
 
 export const GET_STREAMS_BY_GROUP_ID_QUERY = gql`
   query getStreamsByGroupId($groupId: String!) {
