+++ src/queries/tasks/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Task-related GraphQL operations exports
+ */
+
+export * from './queries.js';
+export * from './mutations.js';
