+++ src/queries/tasks/queries.ts
@@ -8,7 +8,7 @@ import {
   TASK_ACTUAL_TIME_FRAGMENT,
   TASK_SCHEDULED_TIME_FRAGMENT,
   TASK_INTEGRATION_FRAGMENT,
-} from './fragments/index.js';
+} from '../fragments/index.js';
 
 export const GET_TASKS_BY_DAY_QUERY = gql`
   query getTasksByDay($input: GetTasksByDayInput!) {
