+++ src/queries/user/index.ts
@@ -1,0 +1,5 @@
+/**
+ * User-related GraphQL operations exports
+ */
+
+export * from './queries.js';
